# apps/web/openapi/petclinic-api.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Petclinic MSS301 — Aggregated API",
    "version": "1.0.0",
    "description": "Aggregated from: auth, customers, vets, visits, workflow, billing, products, reviews"
  },
  "servers": [
    {
      "url": "http://localhost:8180",
      "description": "API Gateway"
    }
  ],
...
  "paths": {
...
    "/api/v1/vets/{vetId}/photo": {
      "get": {
        "tags": ["Vet Photo"],
        "summary": "Get photo metadata + presigned URL",
        "description": "Trả 404 nếu vet chưa upload photo. Presigned URL TTL config qua petclinic.files.presigned-ttl (default 1h).",
        "operationId": "getVetPhoto",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VetPhotoResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Vet Photo"],
        "summary": "Upload or replace vet photo (idempotent)",
        "description": "Multipart field 'file'. Max 10MB (config). Content-type: image/jpeg|png|webp. File quá lớn → 400 error.file-too-large. Type khác → 400 error.unsupported-media. Re-upload overwrite cùng object key qua files-service + upsert DB row.",
        "operationId": "uploadVetPhoto",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["file"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VetPhotoResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Vet Photo"],
        "summary": "Delete vet photo (xoá object qua files-service trước, DB row sau)",
        "operationId": "deleteVetPhoto",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/vets/{vetId}/photo/reject": {
      "post": {
        "tags": ["Vet Photo"],
        "summary": "Reject vet photo — STAFF/ADMIN only",
        "description": "Body: { reason: '...' }. Đổi status → REJECTED. Photo ẩn khỏi public.",
        "operationId": "rejectVetPhoto",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": {
                  "type": "string"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VetPhotoResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vets/{vetId}/photo/approve": {
      "post": {
        "tags": ["Vet Photo"],
        "summary": "Approve vet photo — STAFF/ADMIN only",
        "description": "Đổi status PENDING/REJECTED → APPROVED. Photo hiển thị public.",
        "operationId": "approveVetPhoto",
        "parameters": [
          {
            "name": "vetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VetPhotoResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/visits": {
      "get": {
        "tags": ["Visits"],
        "summary": "List visits — USER thấy của mình, STAFF/ADMIN/VET thấy theo filter",
        "operationId": "searchVisits",
        "parameters": [
          {
            "name": "vetId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "petId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["SCHEDULED", "IN_PROGRESS", "COMPLETED", "CANCELLED"]
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageVisitResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Visits"],
        "summary": "Đặt visit cho pet với vet vào giờ xác định",
        "operationId": "bookVisit",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookVisitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/visits/{id}/start": {
      "patch": {
        "tags": ["Visits"],
        "summary": "Mark visit IN_PROGRESS (vet/staff)",
        "operationId": "startVisit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/visits/{id}/complete": {
      "patch": {
        "tags": ["Visits"],
        "summary": "Vet đóng visit + ghi diagnosis/treatment/fee",
        "operationId": "completeVisit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteVisitRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/visits/{id}/cancel": {
      "patch": {
        "tags": ["Visits"],
        "summary": "Cancel visit — owner hoặc STAFF/ADMIN/VET",
        "operationId": "cancelVisit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/visits/{id}": {
      "get": {
        "tags": ["Visits"],
        "summary": "Get visit detail — owner hoặc STAFF/ADMIN/VET",
        "operationId": "getVisit",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VisitResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/visits/availability": {
      "get": {
        "tags": ["Visits"],
        "summary": "Số slot còn trống của vet vào 1 ngày",
        "description": "Trả 12 work-hour 8h-20h với taken/remaining. FE dùng để hiển thị 'Còn X slot' / 'Đã đầy' trên booking calendar. capacity hiện = 2.",
        "operationId": "getSlotAvailability",
        "parameters": [
          {
            "name": "vetId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlotAvailabilityResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/visits/{visitId}/start": {
      "post": {
        "tags": ["Visit Workflows"],
        "summary": "startVisitExam",
        "description": "Bắt đầu khám — bác sĩ xác nhận bắt đầu. Visit chuyển sang IN_PROGRESS.",
        "operationId": "startVisitExam",
        "parameters": [
          {
            "name": "visitId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
    "/api/v1/workflows/visits/{visitId}/complete": {
      "post": {
        "tags": ["Visit Workflows"],
        "summary": "completeVisitExam",
        "description": "Hoàn thành khám — bác sĩ nhập kết quả. Visit chuyển sang COMPLETED.",
        "operationId": "completeVisitExam",
        "parameters": [
          {
            "name": "visitId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteExamRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "PageableObject": {
        "type": "object",
        "properties": {
          "offset": {
            "type": "integer",
            "format": "int64"
          },
          "unpaged": {
            "type": "boolean"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "paged": {
            "type": "boolean"
          },
          "pageNumber": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "SortObject": {
        "type": "object",
        "properties": {
          "empty": {
            "type": "boolean"
          },
          "unsorted": {
            "type": "boolean"
          },
          "sorted": {
            "type": "boolean"
          }
        }
      },
...
      "VetPhotoResponse": {
        "type": "object",
        "properties": {
          "vetId": {
            "type": "integer",
            "format": "int64"
          },
          "contentType": {
            "type": "string"
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "uploadedAt": {
            "type": "string",
            "format": "date-time"
          },
          "presignedUrl": {
            "type": "string",
            "format": "url"
          },
          "status": {
            "type": "string"
          },
          "reviewedBy": {
            "type": "string"
          },
          "reviewedAt": {
            "type": "string",
            "format": "date-time"
          },
          "rejectReason": {
            "type": "string"
          }
        }
      },
...
      "BookVisitRequest": {
        "type": "object",
        "properties": {
          "petId": {
            "type": "integer",
            "format": "int64"
          },
          "vetId": {
            "type": "integer",
            "format": "int64"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "reason": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 0
          }
        },
        "required": ["petId", "scheduledAt", "vetId"]
      },
      "VisitResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "petId": {
            "type": "integer",
            "format": "int64"
          },
          "petName": {
            "type": "string"
          },
          "petBreed": {
            "type": "string"
          },
          "petBirthDate": {
            "type": "string",
            "format": "date"
          },
          "vetId": {
            "type": "integer",
            "format": "int64"
          },
          "customerUserId": {
            "type": "string",
            "format": "uuid"
          },
          "ownerName": {
            "type": "string"
          },
          "ownerPhone": {
            "type": "string"
          },
          "scheduledAt": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string",
            "enum": ["SCHEDULED", "IN_PROGRESS", "COMPLETED", "CANCELLED"]
          },
          "reason": {
            "type": "string"
          },
          "diagnosis": {
            "type": "string"
          },
          "treatment": {
            "type": "string"
          },
          "fee": {
            "type": "number"
          },
          "processInstanceKey": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CompleteVisitRequest": {
        "type": "object",
        "properties": {
          "diagnosis": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 0
          },
          "treatment": {
            "type": "string",
            "maxLength": 4000,
            "minLength": 0
          },
          "fee": {
            "type": "number",
            "minimum": 0
          },
          "serviceProductId": {
            "type": "integer",
            "format": "int64"
          }
        },
        "required": ["diagnosis"]
      },
      "PageVisitResponse": {
        "type": "object",
        "properties": {
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "content": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/VisitResponse"
            }
          },
          "number": {
            "type": "integer",
            "format": "int32"
          },
          "first": {
            "type": "boolean"
          },
          "last": {
            "type": "boolean"
          },
          "numberOfElements": {
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "$ref": "#/components/schemas/SortObject"
          },
          "pageable": {
            "$ref": "#/components/schemas/PageableObject"
          },
          "empty": {
            "type": "boolean"
          }
        }
      },
      "SlotAvailabilityResponse": {
        "type": "object",
        "properties": {
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "slots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SlotInfo"
            }
          }
        }
      },
      "SlotInfo": {
        "type": "object",
        "properties": {
          "workHour": {
            "type": "string"
          },
          "taken": {
            "type": "integer",
            "format": "int32"
          },
          "remaining": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "CompleteExamRequest": {
        "type": "object",
        "properties": {
          "diagnosis": {
            "type": "string",
            "minLength": 1
          },
          "treatment": {
            "type": "string",
            "minLength": 1
          },
          "fee": {
            "type": "number"
          }
        },
        "required": ["diagnosis", "fee", "treatment"]
      },
...
    },
    "securitySchemes": {}
  }
}